requirements.txt
aiogram>=3.13.1
SQLAlchemy>=2.0.35
asyncpg>=0.29.0
python-dotenv>=1.0.1
PaddleOCR>=2.8.1
openai>=1.51.2
gspread>=6.1.2
google-auth>=2.35.0
